requirements.txt
fastapi==0.88.0
uvicorn==0.20.0
appwrite==0.13.0
google-cloud-language==2.3.0
